Consolidated components into a single index file for the time being
